+++ Modules/SWMB/LocalMachine-Uninstall.psm1
@@ -1041,6 +1041,63 @@ Function TweakViewPDFXChange { # RESINFO
 
 ################################################################
 
+# Suppress Skype
+# https://en.wikipedia.org/wiki/Skype
+# | Hive | DisplayName | Publisher | DisplayVersion | KeyProduct | UninstallExe |
+# |:---- |:----------- |:--------- |:-------------- |:---------- |:------------ |
+#  | HKLM | Skype version 8.137 | Skype Technologies S.A. | 8.137 | Skype_is1 | "C:\Program Files (x86)\Microsoft\Skype for Desktop\unins000.exe" |
+#  | HKLM | Skype version 8.138 | Skype Technologies S.A. | 8.138 | Skype_is1 | "C:\Program Files (x86)\Microsoft\Skype for Desktop\unins000.exe" |
+
+# Uninstall
+Function TweakUninstallSkype { # RESINFO
+	Write-Output "Uninstalling software Skype..."
+	$RefName = 'Skype version'
+	@(Get-ChildItem -Recurse 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall';
+	  Get-ChildItem -Recurse "HKLM:\SOFTWARE\WOW6432Node\Microsoft\Windows\CurrentVersion\Uninstall") |
+		ForEach {
+			$Key = $_
+			$App = (Get-ItemProperty -Path $Key.PSPath)
+			$DisplayName  = $App.DisplayName
+			If (!($DisplayName -match $RefName)) { Return }
+			$DisplayVersion = $App.DisplayVersion
+			$KeyProduct = $Key | Split-Path -Leaf
+
+			If ($($App.UninstallString) -match 'MsiExec.exe') {
+				$Exe = 'MsiExec.exe'
+				$Args = '/x "' + $KeyProduct + '" /qn /norestart'
+			} Else {
+				$UninstallSplit = $App.UninstallString -Split "exe"
+				$Exe = $UninstallSplit[0] + 'exe"'
+				$Args = '/VERYSILENT /NORESTART'
+			}
+
+			Write-Output " Uninstalling $DisplayName version $DisplayVersion"
+			Write-Output " Exe: $Exe $Args"
+			SWMB_RunExec -FilePath "$Exe" -ArgumentList "$Args" -Name "$RefName" -Timeout 300
+		}
+}
+
+# View
+Function TweakViewSkype { # RESINFO
+	Write-Output "Viewing software Skype..."
+	$RefName = 'Skype version'
+	@(Get-ChildItem -Recurse 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall';
+	  Get-ChildItem -Recurse "HKLM:\SOFTWARE\WOW6432Node\Microsoft\Windows\CurrentVersion\Uninstall") |
+		ForEach {
+			$Key = $_
+			$App = (Get-ItemProperty -Path $Key.PSPath)
+			$DisplayName  = $App.DisplayName
+			If ($DisplayName -match $RefName) {
+				$DisplayVersion = $App.DisplayVersion
+				$KeyProduct = $Key | Split-Path -Leaf
+				$UninstallString = $App.UninstallString
+				Write-Output " $DisplayName / $DisplayVersion / $KeyProduct / $UninstallString"
+			}
+		}
+}
+
+################################################################
+
 # Dell Appx
 # Uninstall
 Function TweakUninstallDellBuiltInApps { # RESINFO